testing/web-platform/tests/2dcontext/imagebitmap/createImageBitmap-origin.sub.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>createImageBitmap: origin-clean flag</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<script src="/common/namespaces.js"></script>
<script src="/common/canvas-tests.js"></script>
<div id=log></div>
<script>

function assert_origin_unclean(bitmap) {
  const context = document.createElement("canvas").getContext("2d");
  context.drawImage(bitmap, 0, 0);
  assert_throws("SecurityError", () => {
    context.getImageData(0, 0, 1, 1);
  });
}

forEachCanvasSource("http://{{domains[www1]}}:{{ports[http][0]}}",
                    "http://{{domains[]}}:{{ports[http][0]}}",
                    (name, factory) => {
  promise_test(function() {
    return factory().then(createImageBitmap).then(assert_origin_unclean);
  }, name);
});
</script>
